Created basic auth and inter communication, world-server is WIP Added some tests for fbx loading in a dev folder

+++ Server/Zeus.Server.Library.Models/Account.cs
@@ -21,7 +21,7 @@ namespace Zeus.Server.Models {
         }
 
 
-        public int Id { get; protected set; }
+        public long Id { get; protected set; }
 
         public string Login { get; protected set; }
